requirements.txt
numpy
open3d
matplotlib
numpy-stl
numba